validate

+++ conntroller/categoryController.js
@@ -1,4 +1,5 @@
 const Category = require('../models/category');
+const { check, validationResult } = require('express-validator');
 
 
 const getAllCategory = async (req, res) => {
@@ -11,6 +12,15 @@ const getAllCategory = async (req, res) => {
 }
 const createCategory = async (req, res) => {
     try {
+
+        await check('name').notEmpty().withMessage("danh mục không được bỏ trống")
+            .run(req);
+
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() })
+        }
+
         const { name } = req.body;
         if (!name) {
             return res.status(400).json({ message: 'Không hợp lệ' })
@@ -24,6 +34,13 @@ const createCategory = async (req, res) => {
 }
 const updateCategory = async (req, res) => {
     try {
+        await check('name').notEmpty().withMessage("danh mục không được bỏ trống")
+            .run(req);
+
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() })
+        }
         const { name } = req.body;
         if (!name) {
             return res.status(400).json({ message: 'Không hợp lệ' })
@@ -42,6 +59,16 @@ const updateCategory = async (req, res) => {
 }
 const deleteCategory = async (req, res) => {
     try {
+        await check('categoryId')
+            .notEmpty().withMessage('categoryId không được để trống')
+            .isMongoId().withMessage('categoryId phải là một id hợp lệ')
+            .run(req);
+
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
+
         const category = await Category.findByIdAndDelete(req.params.categoryId);
         if (!category) {
             return res.status(404).json({ message: 'Danh mục không tồn tại' });
