+++ conntroller/commentController.js
@@ -1,3 +1,4 @@
+const { check, validationResult } = require('express-validator');
 const Comment = require('../models/comment');
 
 const getAllComments = async (req, res) => {
@@ -9,30 +10,46 @@ const getAllComments = async (req, res) => {
     }
 }
 
+
 const createComment = async (req, res) => {
     try {
-        const { content, post } = req.body;
-        if (!content) {
-            return res.status(400).json({ message: 'vui long nhap ten' })
+        await check('content')
+            .notEmpty().withMessage('Nội dung không được bỏ trống')
+            // .isLength({ max: 100 }).withMessage('Nội dung không được vượt quá 100 ký tự')
+            .run(req);
+
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
         }
+
+        const { content, post } = req.body;
         const comment = new Comment({ content, post });
         await comment.save();
-        res.status(201).json(comment)
+        res.status(201).json(comment);
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
 }
 
+
+
 const updateComment = async (req, res) => {
     try {
-        const { content } = req.body;
-        if (!content) {
-            return res.status(400).json({ message: "Vui lòng cung cấp đủ thông tin bài viết" })
+        await check('content')
+            .notEmpty().withMessage('Nội dung không được bỏ trống')
+            .run(req);
+
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
         }
+
+        const { content } = req.body;
         const comment = await Comment.findByIdAndUpdate(
             req.params.commentId, { content }, { new: true }
         );
-        res.status(201).json(comment)
+        res.status(200).json(comment);
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
@@ -40,6 +57,14 @@ const updateComment = async (req, res) => {
 
 const deleteComment = async (req, res) => {
     try {
+        await check('commentId')
+            .notEmpty().withMessage('commentId không được bỏ trống')
+            .isMongoId().withMessage('commentId phải là 1 id hợp lệ')
+            .run(req)
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
         const commment = await Comment.findByIdAndDelete(req.params.commentId);
         if (!commment) {
             return res.status(404).json({ message: 'Bài viết không tồn tại' });
@@ -49,6 +74,9 @@ const deleteComment = async (req, res) => {
         res.status(500).json({ message: error.message });
     }
 }
+
+
+
 module.exports = {
     getAllComments,
     createComment,
