+++ conntroller/postConntroller.js
@@ -1,3 +1,4 @@
+const { check, validationResult } = require('express-validator');
 const Post = require('../models/post');
 
 
@@ -10,12 +11,19 @@ const getAllPost = async (req, res) => {
     }
 }
 
+
 const createPost = async (req, res) => {
     try {
-        const { title, content, categories } = req.body;
-        if (!title || !content || !categories) {
-            return res.status(400).json({ message: "Vui lòng cung cấp đủ thông tin bài viết" })
+        await check('title').notEmpty().withMessage('Tiêu đề không được bỏ trống').run(req);
+        await check('content').notEmpty().withMessage('Nội dung không được bỏ trống').run(req);
+        // await check('categories').isArray({ min: 1 }).withMessage('Phải chọn ít nhất một danh mục').run(req);
+
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
         }
+
+        const { title, content, categories } = req.body;
         const post = new Post({ title, content, categories });
         await post.save();
         res.status(201).json(post);
@@ -24,8 +32,17 @@ const createPost = async (req, res) => {
     }
 }
 
+
 const updatePost = async (req, res) => {
     try {
+        await check('title').notEmpty().withMessage('Tiêu đề không được bỏ trống').run(req);
+        await check('content').notEmpty().withMessage('Nội dung không được bỏ trống').run(req);
+        // await check('categories').isArray({ min: 1 }).withMessage('Phải chọn ít nhất một danh mục').run(req);
+
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
         const { title, content, categories } = req.body;
         if (!title || !content || !categories) {
             return res.status(400).json({ message: "Vui lòng cung cấp đủ thông tin bài viết" })
@@ -44,6 +61,15 @@ const updatePost = async (req, res) => {
 
 const deletePost = async (req, res) => {
     try {
+        await check('postId')
+            .notEmpty().withMessage('postId không được để trống')
+            .isMongoId().withMessage('postId phải là một id hợp lệ')
+            .run(req);
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
+
         const post = await Post.findByIdAndDelete(req.params.postId);
         if (!post) {
             return res.status(404).json({ message: 'Bài viết không tồn tại' });
